requirements.txt
markdown-pdf
openai
python-dotenv
PyYAML